rename : 동행 package 이름 변경
