--- src/main/java/com/crude/travelcrew/domain/board/contents/CrewGender.java
@@ -1,8 +1,0 @@
-package com.crude.travelcrew.domain.board.contents;
-
-import lombok.Getter;
-
-@Getter
-public enum CrewGender {
-	MALE, FEMALE ;
-}
