--- src/main/java/com/crude/travelcrew/domain/board/contents/CrewStatus.java
@@ -1,8 +1,0 @@
-package com.crude.travelcrew.domain.board.contents;
-
-import lombok.Getter;
-
-@Getter
-public enum CrewStatus {
-	OPEN, CLOSE ;
-}
