+++ src/main/java/com/crude/travelcrew/domain/crew/constants/CrewGender.java
@@ -1,0 +1,8 @@
+package com.crude.travelcrew.domain.crew.constants;
+
+import lombok.Getter;
+
+@Getter
+public enum CrewGender {
+	MALE, FEMALE
+}
