+++ src/main/java/com/crude/travelcrew/domain/crew/constants/CrewStatus.java
@@ -1,0 +1,8 @@
+package com.crude.travelcrew.domain.crew.constants;
+
+import lombok.Getter;
+
+@Getter
+public enum CrewStatus {
+	OPEN, CLOSE
+}
