+++ src/main/java/com/crude/travelcrew/domain/crew/dto/CrewListRes.java
@@ -1,9 +1,9 @@
-package com.crude.travelcrew.domain.board.dto;
+package com.crude.travelcrew.domain.crew.dto;
 
 import java.time.LocalDate;
 
-import com.crude.travelcrew.domain.board.contents.CrewStatus;
-import com.crude.travelcrew.domain.board.entity.Crew;
+import com.crude.travelcrew.domain.crew.constants.CrewStatus;
+import com.crude.travelcrew.domain.crew.entity.Crew;
 import com.crude.travelcrew.domain.member.entity.Member;
 
 import lombok.AllArgsConstructor;
