+++ src/main/java/com/crude/travelcrew/domain/crew/repository/CrewCommentRepository.java
@@ -1,11 +1,11 @@
-package com.crude.travelcrew.domain.board.repository;
+package com.crude.travelcrew.domain.crew.repository;
 
 import java.util.List;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.crude.travelcrew.domain.board.entity.CrewComment;
+import com.crude.travelcrew.domain.crew.entity.CrewComment;
 
 public interface CrewCommentRepository extends JpaRepository<CrewComment, Long> {
 	List<CrewComment> findByCrewId(long crewId, Pageable pageable);
