+++ src/main/java/com/crude/travelcrew/domain/crew/repository/CrewRepository.java
@@ -1,4 +1,4 @@
-package com.crude.travelcrew.domain.board.repository;
+package com.crude.travelcrew.domain.crew.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.crude.travelcrew.domain.board.entity.Crew;
+import com.crude.travelcrew.domain.crew.entity.Crew;
 import com.crude.travelcrew.domain.member.entity.Member;
 
 public interface CrewRepository extends JpaRepository<Crew, Long> {
