+++ src/main/java/com/crude/travelcrew/domain/crew/service/CrewScrapService.java
@@ -1,14 +1,14 @@
-package com.crude.travelcrew.domain.board.service;
+package com.crude.travelcrew.domain.crew.service;
 
 import java.util.Objects;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.crude.travelcrew.domain.board.entity.Crew;
-import com.crude.travelcrew.domain.board.entity.CrewScrap;
-import com.crude.travelcrew.domain.board.repository.CrewRepository;
-import com.crude.travelcrew.domain.board.repository.CrewScrapRepository;
+import com.crude.travelcrew.domain.crew.entity.Crew;
+import com.crude.travelcrew.domain.crew.entity.CrewScrap;
+import com.crude.travelcrew.domain.crew.repository.CrewRepository;
+import com.crude.travelcrew.domain.crew.repository.CrewScrapRepository;
 import com.crude.travelcrew.domain.member.entity.Member;
 import com.crude.travelcrew.domain.member.repository.MemberRepository;
