+++ src/main/java/com/crude/travelcrew/domain/crew/service/MemberListService.java
@@ -1,13 +1,13 @@
-package com.crude.travelcrew.domain.board.service;
+package com.crude.travelcrew.domain.crew.service;
 
 import javax.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.crude.travelcrew.domain.board.dto.MemberListReq;
-import com.crude.travelcrew.domain.board.entity.Crew;
-import com.crude.travelcrew.domain.board.repository.CrewRepository;
+import com.crude.travelcrew.domain.crew.dto.MemberListReq;
+import com.crude.travelcrew.domain.crew.entity.Crew;
+import com.crude.travelcrew.domain.crew.repository.CrewRepository;
 import com.crude.travelcrew.domain.member.entity.Member;
 import com.crude.travelcrew.domain.member.repository.MemberRepository;
